## background
Implementation of `CKB Light Client` on relayer part of Forcerelay

## task
- [x] generate proof while building IBC transaction to Axon
- [x] verify generated proof using `axon-tools`
- [x] write commitment path and value into IBC transaction's witness
field
- [x] update `ibc-ckb-contracts`

+++ tools/ibc-test/src/tests/ckb/packet/utils.rs
@@ -1,4 +1,3 @@
-use ckb_ics_axon::object::State;
 use ckb_jsonrpc_types::{OutputsValidator, TransactionView as JsonTxView};
 use ckb_sdk::constants::TYPE_ID_CODE_HASH;
 use ckb_sdk::rpc::ckb_indexer::ScriptSearchMode;
@@ -13,6 +12,7 @@ use ckb_types::prelude::{Builder, Entity, Pack, Unpack};
 use ckb_types::H256;
 use eyre::{eyre, Result as EyreResult};
 use forcerelay_ckb_sdk::ckb_ics_axon::message::Envelope;
+use forcerelay_ckb_sdk::ckb_ics_axon::object::State;
 use forcerelay_ckb_sdk::ckb_rpc_client::CkbRpcClient as SdkRpcClient;
 use forcerelay_ckb_sdk::config::{AddressOrScript, Config as SdkConfig};
 use forcerelay_ckb_sdk::search::{
@@ -255,7 +255,7 @@ pub fn prepare_artificials(
         .hash_type(ScriptHashType::Type.into())
         .args(CLIENT_TYPE_ARGS.as_bytes().pack())
         .build();
-    let channel_number = u16::from_str(
+    let channel_number = u64::from_str(
         channel_id
             .to_string()
             .split('-')
